// src/types/fetchCampersParams.ts
export interface FetchCampersParams {
    page?: number;
    limit?: number;
    location?: string;
    form?: string;
    bathroom?: boolean;
    AC?: boolean;
    transmission?: string;
    kitchen?: boolean;
    TV?: boolean;
    refrigerator?: boolean;
    microwave?: boolean;
    radio?: boolean;
  }